\name{dotC-package}
\alias{dotC-package}
\docType{package}
\title{
\packageTitle{dotC}
}
\description{
\packageDescription{dotC}
}
\details{

This package is an attempt at learning the .C interface function. I
call a simple C routine.

}
\author{
\packageAuthor{dotC}

Maintainer: \packageMaintainer{dotC}
}

\keyword{ package }

